mengubah home page menjadi di homePage.tsx karena ada searchparams, menambahkan link agar dapat langsung join dan isi nama, menambahkan qr code
